Fix Python 3.9 removal of node.getchildren()
In Python 3.9, node.getchildren() was removed. We can replace the same functionality with list(node) instead (sesrch for `getchildren` in https://docs.python.org/3.9/whatsnew/3.9.html). I’m not a Python expert, but this seems to work and now passes the unit tests in tox, which I’ve updated to include Python 3.9.