textwrap2: Correctly hyphenate overly long words
The case where a word needed to be hyphenated more than once for the
text to fit within the text width was not handled correctly: after the
first hyphenation, the remained would still be longer than the maximum
line length and the remainder would be forcibly broken up; due to
cur_len not being updated when the first part is added to the line, the
resulting line would then be malformed.

With this commit, a chunk is only forcibly broken up when the last chunk
had not been hyphenated. Other solutions, e.g. only breaking a chunk
when it is the first of its line, lead to behaviour for long
un-hyphen-able text inconsistent with the standard library’s textwrap.

A unit test for both a natural word and an unbreakable string too long
to fit on a line is added as well.

Fixes #11.